move map to map folder
